Split Interpreter Index factories to improve compilation time
